disd

+++ THIRDWEEK/8Gorner.py
@@ -2,6 +2,6 @@ n = int(input())
 x = float(input())
 i = 1
 while i <= n + 1:
-    k = int(input())
+    k = float(input())
     i += 1
     print(k)
